renaming journal-components folder to components
