+++ src/pages/journal/__mocks__/journal.module.mock.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { IonicModule } from 'ionic-angular';
-import { TestSlotComponent } from '../journal-components/test-slot/test-slot';
-import { TestDetailsComponent } from '../journal-components/test-details/test-details';
-import { IndicatorsComponent } from '../journal-components/indicators/indicators';
-import { TimeComponent } from '../journal-components/time/time';
-import { CandidateComponent } from '../journal-components/candidate/candidate';
-import { TestOutcomeComponent } from '../journal-components/test-outcome/test-outcome';
+import { TestSlotComponent } from '../components/test-slot/test-slot';
+import { TestDetailsComponent } from '../components/test-details/test-details';
+import { IndicatorsComponent } from '../components/indicators/indicators';
+import { TimeComponent } from '../components/time/time';
+import { CandidateComponent } from '../components/candidate/candidate';
+import { TestOutcomeComponent } from '../components/test-outcome/test-outcome';
 
 @NgModule({
   declarations: [
