+++ src/pages/journal/components/candidate/__tests__/candidate.spec.ts
