+++ src/pages/journal/components/candidate/candidate.html
