+++ src/pages/journal/components/candidate/candidate.scss
