+++ src/pages/journal/components/candidate/candidate.ts
