+++ src/pages/journal/components/indicators/__tests__/indicators.spec.ts
