+++ src/pages/journal/components/indicators/indicators.html
