+++ src/pages/journal/components/indicators/indicators.scss
