+++ src/pages/journal/components/indicators/indicators.ts
