+++ src/pages/journal/components/slot/slot.ts
