+++ src/pages/journal/components/test-details/__tests__/test-details.spec.ts
