+++ src/pages/journal/components/test-details/test-details.scss
