+++ src/pages/journal/components/test-outcome/test-outcome.scss
