+++ src/pages/journal/components/test-slot/test-slot.scss
