+++ src/pages/journal/components/time/time.scss
