+++ src/pages/journal/components/time/time.ts
