+++ src/providers/slot-selector/slot-item.ts
@@ -1,5 +1,5 @@
 import { Type } from '@angular/core';
-import { SlotComponent } from '../../pages/journal/journal-components/slot/slot';
+import { SlotComponent } from '../../pages/journal/components/slot/slot';
 
 export class SlotItem {
   constructor(public component: Type<SlotComponent>, public slotData: any) { }
